Added Pascaline identifier goto labels
1. Label entries get a label string as well as value. The string
is a quanta, its base pointer flags if it is a string or numeric
label. Note that P6 keeps its own internal sequential numbering
system for labels.
2. Change syntax to allow this unconditionally, but error on
non-standard use. Also error changes to new "integer or id"
error on Pascaline.
3. searchlabel gets an id mode flag and now searches for both
label types.
4. newlabel gets a mod flag.
5. labeldeclaration and gotostatement are fairly easy.
6. Statement is the worst:

mylabel: if then else

Could easily be:

mylabel := 1;

With an undefined label. We instituted a little look ahead
mechanisim that saves the next symbol and walks forward looking
for the ':' terminator. This is pretty light weight because
we only have to save the id if it is indent. It has to flag the
next statement for assign/call because insymbol might be standing
on the tolken just after the ident.

This could be converted into a general purpose lookahead, but I
don't believe it is needed again. This, by the way, is why N. Wirth
specified numeric labels (or at least one reason).