Pascaline extended file procedures
This is a skeletal commit. The support is not complete, and there
are known issues in the rest of the code as well.

1. Increased number of standard identifiers, and filled out the
table with new procedures, functions and header file ids. Note
that a few of the range ids used in place are now formal types.
2. Create a series of dummy routines to handle new procedures/
functions. Linked to standard proc/func call mechanisim.
3. Fixed bug with label handling in statement. The lookahead was
not correct.
4. Added shorthand procedure entstdprocfunc to entstdnames. This
dramatically compresses the code needed to create standard
procedures and functions. Used to create new proc/func entrys
as well as clean up the previous ones as applicable.

Next is regress bugs, fill out new procedures.