Added GPC file handling definitions
1. Created filsts definition.
2. Struck equivalence notations. Tried it, does not work (changes
the state of the file, which is not acceptable). Oh well.
3. Various reformat.
4. Various routines filled out with GPC definitions.
5. Added error for position < 1.
6. Changed Makefile to allow use of GPC extensions.

There's a lot missing from GPC file handling. The GPC docs are
in a very poor state, and some of what was in the docs does not
work.

I band-aided the make process to get this to run. A better
solution will come later.