Added halt procedure
1. Reverted GPC file support to commented out.
2. Back to ISO 7185 mode on GPC.
3. Halt support front end, back end. Simply aborts the run.

Why back to ISO 7185 support? GPC shows "fall off" type behavior,
which I covered in the Pascal FAQ. That is, they got compliance
by changing the basic behavior of the code when the ISO 7185
switch is on, vs. when it is off. This means that if you want
to use the extensions in the host compiler, you may have to fix
issues in the code or it may not compile at all, due to the
non-compliance with ISO 7185 now in the compiler with the ISO 7185
switch off.

GPC was not bad in this respect (the EOF handling changes), but
that makes me nervious. Better to do development with the switch
on.