Anonymous function results and declaration order relax
1. Added inner loop to block declarations that continues if in pascaline
mode and further declarations exist.

2. Added errors for conflicting function results.

3. Removed old commented out code from P4.

4. Added skip flag to selector() and assignment() to change their
modes so that they can process the "result x" statement. This
basically short circuits their parsing.

5. Added "result" symbol to symbol checks.

6. Added result handler pseudo-statement, with various error
checks.

I actually did try a version with factoring out the needed code
from assignment() and selector() into the main code, but the
result was much larger and uglier. Using the assignment() and
selector() procedures is not clean, but it is way more compact
because all of the handler code is already there.