Second part of constant expressions
Changed real input system from ascii representation to binary.

1. insymbol now parses numbers entirely in binary. Reals
start as integer, then the mantissa gets converted to the
whole part of a real and scaled.

2. All references to reals in the code changed to reflect
this.

3. Old systems to handle reals as strings removed.

Ok, so why this change. Wirth's old ascii based system
was good in that it didn't convert to binary, which then
gets converted to ascii for the intermediate, then again converted to
binary. This opens up the possibility of convertion loss, as well as
being far less machine independent.

However, this was a cumbersom system, and with constant expressions
would get worse. Further, in constant expressions it would have
converted to and from the binary format, multiple times in fact,
so the problem would have been worse. In practice the binary
system works quite well.

Note that a new limitation was introduced, which is that the
whole part of a constant real is limited to the size of integer.
For contrast, see "readr" in the backend, which does the
calculation entirely in real mode.