Boolean integer operations
1. Added xor mach instruction.

2. Fixed old mest() error that had the wrong check for bad values.

3. Fixed error with 80 (*ckl*) instruction. See pascal-p5 bug ticket

3. Enumerated the 400 and 500 errors. I was tired of always seeing the
same 500 number on all compiler errors, so now all different but
grouped into 400 and 500's, meaning compiler error.

4. Changed "not" to accept both pairs of bool and int, then moved it to
the typed gen routine. That means that not, uniquely, will have integer
and boolean equivalents. for extra credit tell us why before you get to
the end of this note!

5. Changed "and" routine, but generates only one instruction. Same
with "or", and added "xor" as a copy of or.

6. Added boolean integer emulation routines in pint (same code as pcom).

7. Changed "not" instruction to have notb and noti variants. Added
xor instruction.

8. noti, and, ior and xor instructions get (re)done in terms of
emulation routines. notb stays at the old method.

So why is not the exception to only one logical bit instruction has
to consider its boolean/integer status, not?

Boolean is 0 and 1. and, or, and xor do the right thing even applied
to all bits. Bit one being 0 or 1 will result in the same bit with
others zero on a true boolean. However, not flips all bits. The result
is that we have to treat boolean not specially.