Automate changeover for language extension routines
1. Changed the format for the entry/exit comments to one more
acceptable to sed.

2. Created the scripts to change the host compiler.

The scripts now give the ability to change the host compiler by
switching the part of the extention routines in pint.pas to
uncover the code for a particular host. This works by having
a lead in comment and a lead out comment for the particular host,
then that is changed so that it either comments out from the
lead in to the lead out, or just comments out the lead in and
lead out, thus uncovering the host code contained.

These scripts are available:

setgpc.bat - sets the extention routines to GPC host.
setiso7185.bat - sets for ISO 7185 mode, which is no extension
support - calling the routines gives an error.