View and out parameters
1. Flip view parameters to formal (var) mode if the type is strutured.

This is the third and final stage of view and out parameters.
Wirth & CO. really made this easy, and its a textbook implementation
of the concept. Things loaded onto the stack become parameters.
If it is structured (> power) then it gets an address, otherwise
loaded to stack in whole (even if it is a set with 32 bytes).
When the target procedure is called, it can either use the pointer
in the stacked parameter or make a copy in the local frame and use
that. Its all keyed to the actual/formal status of the parameter.
So by creating a view parameter as actual if <= power, and otherwise
formal, we show why view parameters are such a good deal. Small
parameters get copied to the stack, but big ones stay in place.
Its essentially selective application of var vs. value parameters.