Array type shorthand
1. Modified array handler in typ() to alternatively process Pascaline
array types. Because a missed '[' leader in ISO 7185 could be a mistake
and would cause a significant mistaken parse if we process that as
Pascaline, I head into the ISO 7185 parse even if the '[' is not seen
and we are in ISO 7185 mode.

Pascaline array shorthand is fairly simple. It basically uses the array
index whole size to contruct a subrange of 1..n where N is the size of
the index. Thus internally the forms:

array 10,10 of ...

and

array [1..10,1..10] of ...

is identical. Thus, it is just a shorthand.