Extended write/writeln statements
1. Added new operators to write integers in hex, octal and binary.
These are selected by $, &, and % characters in writeprocedure.

2. Added same to backend. writei() is extensively redone to accept a
radix parameter.

So this change is local to P6 (not part of Pascaline). It is mentioned
in the Pascaline rationale, though.

A write parameter can be terminated by a $, & or % (hex, octal or
binary) symbol:

write('The number is: ', 100$:10);

It means to format the output in that radix.