Extended write/writeln statements
1. Added new "padded string" field macro.

This is unique to P6. You can specify:

write('hi there bob       ':*);

and it means only print the contents of a right padded string, IE,
remove all right padding before print. This feature was to make up for
the Pascaline change that removed fields of 0 to mean padded strings.

This is nowhere in the Pascaline specification.