Extended write/writeln statements
1. Added zero fill to write statements. A new single character symbol,
'#', was added, and three new variations of write integer. Parsed
'#' symbol at the head of field.

2. Added zero fill operators and handling to back end.

Another (and hopefully the last) off-Pascaline addition to P6, we
can specify zero filled numbers as:

write(42:#10);

The '#' character specifies that the number will be zero filled to the
left. A field width follows as per usual.

Note that left justify using zero fill are no-ops, since only leading
zeros can be added.

This brings P6 to about the same level as C for output formatting.
Note for more advanced formatting, the string library features extensive
formatting functions that are more powerful than anything in the
standard P6 write functions.