Extended read/readln statements
1. Minor change to rationale.

2. Added fielded versions of integer, real, character and string
read routines.

3. Changed readprocedure() to parse field, padded string marker,
and generate new fielded routine versions.

4. Add new instructions to backend, with new handling.

5. Extensive changes to read routines to handle fielding.

All of the input convertion routines now handle fields. The default
setting for the field width is maxint, so that effectively the field
width is "infininite". However, this is not sufficient to handle
both fielded and non-fielded cases, because in cases the handling
of the two is different. For example, the default field handling
is to check all trailing characters in a field for space, and
obviously you don't want an infinite field checked for spaces.
Thus the routines accept both a field width and an on/off flag.