Redeclaration of forwarded procedures and functions
1. Added new routine to recycle linear id list.

2. Error for repeated parameter lists now conditional on iso7185.

3. parameterlist() throws a new level if it is parsing a duplicate
parameter list, so that the names in the list can be thrown away.

4. After parameterlist() call, procdeclaration() recycles the parameter
list given it by parameter list if it's a duplicated forward.

This is the first stage of forward procedure/func relists. It tolerates
the relist, but does not check it. The next stage is to check it.

Also we need to check a result type exists on dups.