Added program header standard bindings
1. Added three new file definitions, error, list and command. Refactored
the init of files into routine, and put all file inits there.

2. Expanded the check for valid external files.

3. Added file numbers to pint, and to valfil().

4. Added low level primitives for read functions using file numbers.
Refactored read routines to use that.

5. Refactored csps to use new method.

So the need to expose a general purpose interface to implement files
kicked off an extensive refactor of the code.  The situation is that a
"fileofiled" variable like the command file, which is satisfied by a
command line, must connect to all of the normal read procedures (text
only).

The answer was to stop passing files down to the read routines, and
instead pass down the logical file numbers. These are then send down
again to a series of small routines that implement the file primitives
like read buffer variable, get(), eoln(), eof(), etc.

This means we can implement our own primities, and it also had the side
effect of dramatically simplifying the csp handlers. I may, in fact, use
the system on writes as well and gain the advantage of simplifying the
code further.

At this stage, error and list files are implemented as aliases to the
output file. The command file is stubbed out. Also need a check for if
the file used actually appears in the header.