Fix problems with header file verification
There were major holes in the verification of header files. The prd and
prr files were not checked to be in header if they were used, and the
check if the input and output file were in the header was easily
bypassed if the file was explicitly used, IE., write(output, ...);

1. Moved the "appears in header" flag into ctp entries.

2. Added checks for all files and the new files added with Pascaline.

Note that this patch could be used to retroactively fix P5 as well.