Undefined program parameter binding
1. Minor document correction.

2. Added two intermediates to connect external header files.

3. Moved header file handling to it's own function.

4. Added generation of external header file connectors.

5. Added new backend routines to support header file connections. These work
by reading from the command file, so this should not need new language extension
support other than what was previously provided.

The header files are satisfied, in appearance order, from the command file and
thus from the program command line. These will all be satisfied before the
program runs. The command file will be left just after the filenames are
parsed, so further parsing can be done on the command line after the
filenames appear.

The external name assignment also carries the original header names down to
the bottom routines, even though we don't use them. This allows the implementor
to use other methods to satisfy the header names. for example, the GPC method is
to connect them to files with the name of the header. That is easy to do.