Character escapes
1. Added escchr routine to search and process character escapes.

This is part one of character escapes. In the second part will add numeric
character escapes.

The search used is clearly braindead (linear search), but it only happens if
there is an escape leader ('\'). Thus I would argue not so bad.

Note in the code that we use the ASCII code for the escape leader character
(92). This means that the code is compatible between both ISO 7185 and
Pascaline modes. Otherwise it would be '\' in ISO 7185, and '\\' in Pascaline.