Character escapes
1. Added numeric force escapes to string input.

2. Fixed P2 and Pascal-s source to be standard agnostic by removing '\'
references. They were being used to flag as error characters, and there are
plenty of other characters that don't get flagged in any case, so no loss.
Note that P4 didn't have that.

3. Startrek got changed to IP Pascal long ago, and I don't feel like changing
it to be agnostic. So fixed it back to ISO 7185 and then put an ISO 7185 flag
at the top, which makes it a good test of the flag. Fixed the compare file
to match.

This completes the character escapes feature. Note that it only works on strings
in the buffer, which means that tricks like 'don\'t stop' don't work, because it
does not change the way quotes work (for example IP Pascal will do that, and
escape line endings as well). 'don''t stop' still works, and is the right way
to do that.

This is a prime example of why character escapes is an Annex, not in the main
standard.