Extended case statements
This patch adds fallback code generation for overly sparse case tables.

1. Added new instruction, cjp. This instruction compares for a range of numbers,
then pops the selector and goes to the target label if the compare is true. This
replaces a lot of instructions if the operation were carried out using normal
instructions.

2. Added occupancy threshold equate.

3. Added generator for cjp. cjp is like a cross between chki and tjp.

4. Added routine to count active cases in case list.

5. Added occupancy check. Less than the threshold (currently set to %50
occupancy), we devolve to using a series of compare/jumps using cjp.

6. Removed extraneous debug messages.

7. Fixed no frame handling in errorm (missed exception handling).

8. Added cjp to backend. We are reusing old instruction slots now, because we
may run low on instructions at some point.

9. Added generation and handling for cjp instruction.

This completes extended case statement handling. A remaining question is what
purpose the general case table size limit serves. It is presently 10,000. This
would be difficult to generate source-wise (although as noted P5 could actually
generate an overflow as well).