Debug mode - symbols passthrough
Added block marking, symbols output and symbols type digests.

Each block is marked with begin and end markers in intermediate. The first
letter indicates the operation:

b <type> <name> - begin block
e <type> - end block

The type is:

p - program
r - procedure (routine)
f - function

The name is the name of the program, procedure or function.

Symbols have the format:

s <name> <type> <address> <digest>

The name is the name of the symbol, without qualident. The client must figure
out the block nesting itself from the begin and end marks.

The type is g or l. g for global, l for local.

The address is an offset, either from the globals area base, or from the frame
for locals.

The digest or type digest (or "type spagetti) is a shorthand description of the
type of the variable. The nesting structure of the type is listed from most
complex to least complex. Each type element is represented by a single letter.
The type elements are:

i - integer.
b - boolean.
c - character.
n - real (number).
x(<val>[,<val>]...)<base> - Subrange or enum, with the values enclosed, followed
 by the base type.
p<base> - pointer, followed by base type.
s<base> - set, followed by base type.
a<index><base> - Array, followed by index type, then base type.
r(<field>[,<field>]...) - record, with list of fields. If a variant is present,
 the tagfield will have a sublist like: t(0(a, b, c)1(d, e)). The list represents
 each variant. Each field has the form: <name>:<offset>:<type>. It contains the
 name of the field, the offset from the record base address, then the type digest
 for that. The name of the tagfield may be omitted for undiscriminated tagfields.
f<base> - file followed by base type.
e - exception.

If any type is not defined, it shows as '?'. Usually this shows the result of
a compiler error (and resultant mangled type). If it shows up without a
compiler error it is a malfunction and should be corrected.

Types in the digest are complete. There is no facillity for named types. This
means that all types in the digest are complete, even if the format is somewhat
redundant with other symbols.

Pointers create cycles in the type description, and that creates a problem with
pointer type representation. I plan a fix for this, but in the meantime all
pointers give an unknown type ('?'). The plan is to recognize cycles in the
type and output a number for the type instead of the type digest, which would
mean "repeat at this offset from the start of the digest", so:

pi = pointer to integer.
p5 = Cycle back to the 5 character position of the digest.

In the back end, all of the symbols and block markers are treated as comments
at the moment. They will be used in further patches.