Debug mode
1. Added block/symbol tracking system.

2. Increased file string size to accomodate digests. These can get fairly large.
Unfortunately this is all jamed into the file string format, which perhaps
should be separated out.

3. Added a fixed to variable string routine to take filenames/symbols/digests.

4. Changed dd command to ignore marks that are not activated.

5. Added a test command to dump blocks and symbols. The user does not need this,
is a fixer command only.

This completes intake of symbols into a block structured table. We have an
algorithim to tell if the symbol is active, which I will detail in the next
patch. This then forms the basis for the symbols commands.