Debug mode
1. Changed spacing in set output (ldcs). It was verryyy wide, and turns out I
did that a long time ago. Oh well.

2. Added break table index types.

3. Added lower case convertion and variable length to fixed compare for symbols.

4. Added isbrkl() to check if breakpoint lies on the given line.

5. Removed dmpins, debug mode replaces this functionality.

6. Added getlab() routine to load, used that to simplify symbol and type reading
from intermediate.

7. Changed source listing to use the line breakpoint sets and not machine code.
See other comments below.

8. Changed getchr to nxtchr in debug parsers, it was conflicting with getchr()
store pickup routine.

9. Added ability to search symbol through machine frames, ie., search a given
symbol to see if it lives in the active frames, then returns what frame and
what symbol entry.

10. Generally filled out prttyp(), not done yet. It does subranges, enumerations
and improved array processing.

11. Fixed so that when breaking on a breakpoint that has a line assocated,
srclin is set to that. See below comments.

12. Added code to walk forward over source line markers. See comments below.

13. Filled out p(print) command.

14. Fixed bug on first code breakpoint, source line not set.

Two changes were made with source line handling around breakpoints that improve
the user experience. First, we walk forward over any source line marker
instructions until we find real code. This was because there are a lot of lines
that just generate markers that never really get executed. Now we assume that
for any given source, breaking on a line means to break on the first code
following that.

Second, we keep what line we used for a breakpoint, and show that in listings,
as well as fix the source line pointer to reflect that. The reason for this
is that several lines can refer to the same bit of code. What we do now is take
the users word for it as to what line is meant. IE., if lines 2, 3 and 4 all
end up at the entry for a function, if the user says "break at 3", we show
the break at 3.