Debug mode
Added watch variable command.

1. Added watch tables and control variables.

2. Added iswatch() to check if a watch is set at given address.

3. Store instructions in interpreter changed to post watched changes to the
debugger.

4. Added handler for the watch to the debugger proper.

5. Added commands to set, clear and list the watch table.

So a brief overview of how it all works. The watch addresses are stored in a
table of up to 10 watchpoints. The addresses are of variables in the program,
and only simple scalar variables are supported, since we only check writes
to the base address. A structured variable may or may not write its base
address, so it would have to be a range of addresses, which is a lot more
expensive to match for. At 10 simple addresses, the overhead to match any
of those addresses is limited, and in any case only happens when the address
gets stored to.

Besides the watch address, another table, wthsym, keeps the symbol being watched
as well as the subtype that is being watched. This tells debug how to print the
contents out. The tables are kept separate so that the watch address table is
very simple and even stupid compilers can generate an effective table search
for it.

The procedure to watch any address is done by leaving the variable "stopwatch"
set, which causes the store handler to post its address to store in stoad and
then stop the instruction in a "rerunable" state, that is, with the stack left
in the same state so that the instruction can be executed again. The watchmatch
variable is set to indicate why the instruction stopped. The debugger prints
the contents of the watched variable, then reruns the instruction with
"stopwatch" turned off, then prints the variable contents after the instruction
has modified it. Then stopwatch is turned back on and the run continues.

In this way every modification to the watch variable is caught. The normal
(non-watched) instruction path is to do a table search on the watch address,
then early exit the instruction if that is found. The biggest overhead, to bail
out of the instruction, enter the debugger and print variable contents, only
happens on watched variables, so the normal program slowdown is minimized.

I will effect a further change in the next patch that only searches the watch
table over active watch addresses, and hopefully further speed the algorithim.