1. Reverted iso7185pat test to not list.
2. Fixed issue with stop and break where the pc was getting set indeterminately
after. A stop or a break should leave the PC at instruction start.

3. In connection with that, modified singleins() so that each of the trapping
instructions leave the pc at the instruction start. This is the way a real
CPU works, and is a clean solution to traps in general. It also removes the
need to fix up the pc in debug mode, and the need to keep startins as a global.
I reverted the start address save to pcs.

4. Fixed bug in dd debugger command.