Added cycle detection for pointers in type digest
1. Added new error for type tracking array overflow (too complex).

2. Reworked output from wrttyp() subroutines to count outputs.

3. Added tracking array for stp types. Only the lead stps are stored in the
array, other things like fields, cannot be directly pointer referenced.

4. Pointer generation now searches backwards for a type match, then outputs a
backward reference if a cycle is detected.

5. Added a resolve step just after types are parsed for vars. See below.

So the problem being addressed is cycles in pointers. A pointer may or may not
have a cycle, and having a circular pointer reference is not fatal to the
debugger. Having an expression like:

type recp = ^rec;
     rec  = record i: integer; next: ^rec end;

var r: recp;

r^^^^^...

Simply means you are walking a list. You can do that for as long as you have
characters in the command line. However, trying to generate that would lock up
the compiler. So we record each stp type that occurs on a "node", of a type,
that is, a whole type and not a record component, in an array according to
the position in the type digest.

We search that array, and if found, we output a number signifying it's position
in the digest. So to resolve that type, the parse for pointer type simply goes
to that looped position in the digest, and so no issue.

If the pointer type does not have a cycle, then it is output as any other
subtype, of any complexity.

The need to resolve forward reference pointers (resolvep()), comes about because
the compiler chooses to treat all pointers as forward references, then resolve
them as a group. This causes problems if they are unfinished when we output them
as symbols in the digest. We solve that by resolving after each type parse in
vardeclaration. This is not as bad as it looks, since it simply empties out the
forwarding list. Its reasonable because there is nothing to forward to. There
are no more type definitions. You are simply resolving types that by definition
already have (or will never have) forwards.