Added tracking of on new line/not on newline output. This is used to normalize debug output and make sure we don't print in the middle of a line.
1. Added newline variable.

2. writelns to any "output" or alias set newline true, all other outputs set
newline false. page(output) is assumed to also set newline.

3. In one case, renamed a local named "newline" to something else to avoid
confusion.

4. Created wrtnewline routine that will make sure we are on newline.

5. prefaced all debug outputs with wrtnewline. Error handling will be redone in
the next patch.

Background: debug assumes that the program may have left the output print
position in any particular state, and so usually kicks a line, both to prevent
midline printing, but also to create space between user prints and debug prints.
This solution works, but produces random spacing since we don't know where the
printout position for output is. This solution fixes that. All we need to know
is if the output is on a new line or not. Then if not, wrtnewline kicks it.

Of course the "delux" algorithim is the (what I call) Univac 1103 solution,
where you buffer the contents of output lines, and if there is another task,
such as debug, you return the line, overprint the output, then spill the output
line later when the task finishes (the debugger resumes).

This is a good example of the fact that we still have things to learn from
the mainframe era.