Unified all debug errors in one routine
1. Implemented error() routine in debug.

2. Collected all errors into that.

3. Added wrtnewline() calls into stop/break messages in main.

How newlines work with debug:

As stated in the last patch, we sync up the newline status when debug mode is
entered. Debug itself is "self justifying", that is, each command or action is,
or is supposed to be, started at on a newline, and ends with a newline. Further,
most debug commands that print kick a blank line before the data they print and
then again after.

Thus: error() will call wrtnewline(), but in virtually all cases, it will do
nothing. This is because debug mode kicks a line, it debug itself does not track
line endings, but rather relies on it being self consistent. Thus the newline
flag will stay on during the entire run of debug mode, and the wrtnewline()
at the start of error() does nothing.