Debug mode
Added formatting for structured types in p (print).

1. Removed extraneous print from skptyp().

2. Refactored subrange prints in prttyp() using getrng().

3. Several prttyp terminals were not treating their base types properly.

4. Added indentation for complex types.

5. Added complex(), a routine that tours a fieldlist to N depth and finds
substructures.

6. prttyp(), array and record prints check for substructure and flip to indented
mode if found.

7. All calls to prttyp() refactored to take indent.

The new rules for printing work like this. All items, even arrays and records,
are normally printed on one line, separated with spaces. If there is any
substructure to a structured type, then each component is printed on its own
line. Note that a substructure is a set, array or record, even though sets
themselves cannot contain substructure. This rule has non-obvious implications.
For example, if ANY field of a record, even nested within other records, is
itself complex, then the WHOLE record has its components printed one per line.
Note that a record without complex components can be printed on one line,
even though it is itself being printed on an individual line as part of a
record.

In indented mode, each substructure is indented for a new level, which greatly
aids the user to recognize them.