Debug mode
The way we generate line numbers in intermediate was changed. We hold line
numbering until insymbol() is called, then all held lines are output. This
has the effect of pinning all line entries to the following symbol. This greatly
improves the correstpondence of source lines to output code, see the ticket for
examples.

1. Added lineout variable, which tracks linecount.

2. Removed line marker output from endofline().

3. Added new routine, outline() that dumps all of the line markers between the
value of lineout and linecount.

4. Called outline() at the start of insymbol.

5. The final call to outline at the end of the file insures we get all lines in
the file; actually may not be required.