Move constants deck down with program code
The constants deck must be created at top of memory. Basically, the code is
filled from the bottom to the top of memory, and the constants from top to
bottom with an error if they cross. We are moving to a single, unified deck for
several reasons. We do this by holding a directory of relocation pointers for
constant references, then moving the constants deck down over code and
relocating it.

1. Added new "maxtop" in addition to maxstr, because we need both the maximum
memory address as well as +1. This is because one past the last valid address
is a valid address for things like the stack pointer.

2. Added a constants reference directory.

3. Added a new "cststr" to indicate start of contants. The start of constants
address no longer has program function (it used to indicate top of stack). But I
want to report it in diagnostics. It replaces cp which is non-memnomic and I
wanted to be sure all global references to cp are gone.

4. Added alignuc() or "align up with clear". This routine not only aligns, but
clears the skipped space. We use it to generate a clean coordinate deck.

5. Created putcstfix() to add entries to the constants relocate directory.

6. final deck preparation in load() was extensively refactored. See memory map.

This is the first part of the unified deck fix. The second, to follow, is global
variable relocs (get rid of the pctop relative operations). This will allow us
to export a deck without external parameters.