Output unified code deck
Pint now has option to output unified code deck in modified Intel format. See
below.

pint had changes from last commit that apparently got missed. See last commmit
for some of the changes.

1. Added lnp to instructions. Extended startup length.

2. Added file parameter to wrtnum() and wrthex(), necessary because we now use
these routines to create an output hex file.

3. Added "e" output deck option and variable. Locked out other printing
functions. See note below.

4. Added wrtdck() procedure (write deck) that outputs the entire code block to
the prr file.

5. On option, output the deck and stop.

The output deck format is a series of 16 byte lines followed by a possible
partial line and then the end mark. The format of each line is:
:LLAAAAAAAAAAAAAAAADDDDDD...DDCC

Where:

LL    - Length of data, from 1 to 16.
AA... - Address, 64 bits.
DD... - Data, 1-16 bytes.
CC... - Lost carry byte checksum.

The end marker is zero length and zero address, thus:

:00000000000000000000

The format is like, but not equal to Intel hex. We use 64 bit addressing and
don't include a record type. This last is on purpose. I don't appreciate
"kitchen sink" loader formats that try to do too much and want to discourage
that. With that same idea, the address is meant for checking only, and not to
use for (so called) scatter loading. The code is always presented as a
contiguous block (and that can be used to verify and produce error in the
client).

Point #3 above is wrong and will be corrected in the next patch.