Added mach interpreter and friends
1.Makefile was changed to compile pmach.

2. Created new files, compilem.bat, runm.bat, testprogm.bat and regressm.bat.
These do the same thing as the normal versions, but use pmach to run the
program.

3. Changed diffnole.bat to ignore difference between pint messages and pmach
messages so that the same compare files can be used.

4. Checked in the first version of pmach.pas.

Ok, so what is pmach? In a previous patch, we gave pint the capability to output
a binary deck of its code. pmach loads that file and runs it. pmach is a cut
down version of pint, a bit less than half the size. However it runs complete P6
programs. pmach is a response to users who just want to implement the lower
level interpretation without needing to assemble the intermediate. This would
apply, for example, to people making interpreters in C.