Reversed earlier decision to use .pint for intermediate
The old method was to type the intermediate by version, ie., .p5 for the
intermediate. I decided to go back to that method, because we mix a lot of
different versions in various places and that can be confusing. For example,
running p4 on p6 has two different intermediates in the same place.

For similar reasons, I switched to using .p6o for the mach object. I didn't use
.p6m because the object format is not particularly mach specific.