Refactor for code generation at program level
The next series of mods are going to add code generation at the program level.
The way it works currently is that all code generation is enclosed in body().
That makes sense, but results in distortions like the multipass arrangement
to generate the startup code. This gets worse with modules, so time to fix it.

1. Moved putlabel() to join its family.

2. Moved the majority of the gen routines to above block(). The ones that were
not moved were tied into other parts of the code.

3. Moved constant holding table and stack tracking to main.

4. Added init of constant holding and stack track to programme(). Above body(),
we don't do expression/statement generation, its all controls generation (jumps,
stack setup, block setup, etc.). However, the gen routines use these so they
need init.