Refactored out the dual pass in pint
As mentioned, the dual pass thing in pint where it reserves a short section of
code at the start, then backs up and finishes it, was (apparently) just to keep
code generation confined to the block() routine. The last refactor moved the
simple generators up to the main level. This refactor gets rid of the double
pass.

This fix matters to implementing modularity.

1. removed begincode and comment.

2. Moved entname, the label for main program code, to the top.

3. Changed generation of label location in body.

4. Moved startup code generation to programme().

5. Removed second pass from pint.