Extensive refactor to prepare for modularity
In modules, there can be a block, defined as an addressable begin/end block
with framing, separate from the declarations. This used to be bundled in
block(), which contained declarations plus body(). I seaprated off declarations
into declare(), and then body() is removed and stands alone as body(). There are
a few subdeclarations like constant() that had to be moved out and renamed.
Declare() and body() reference each other (of course).

This enables modules to treat declarations and blocks separately, as will be
seen in coming patches.

A PS:

I have seen pint fault at assembly level in sucessive runs of iso7185pat. I was
unable to track this down, being rare and intermittent, but it bears watching.