Modularity
Fixed symbols to be correct for new modular layout.

1. Added gblrlc(), one version that offsets all globals and one that only
offsets the ones in the current module.

2. 'g' instruction offsets all globals by current module base, then sets a new
base from the size in the instruction. Also calculates the total global size.

3. Load() uses gblrlc(off) version to offset all globals to globals area.

4. Added to help menu for debug.

To do: need to remove exception vectors and files from suceeding global modules,
so that only the program module has them.

Continued from last commit: I see I never added the end note for labels, so here
it is:

The label numbers for each module are unique and don't cross from module to
module. Thus the label table only needs to consider the current module. It is
cleared each time a module is finished and it has been used for any relocations.
The module coining of such labels is done mainly for readability. It is stripped
off labels and discarded.

Note that the module coin name could be checked against the current module, and
that may happen in the future.

Note that labels are impossible for the code to programmatically generate, since
the name cannot start with a number, even if it is a numeric goto:

mod.1